stereo: Finish implementing seed_option 0
This allows the user to shut off the calculation of low res
disparity. It's not helpful when the images are smaller than 2000px
diagonally.

+++ src/asp/Core/StereoSettings.h
@@ -44,7 +44,7 @@ public:
   float slogW;                 // Preprocessing filter width
 
   // Correlation Options
-  vw::uint8  seed_option;      /* 0 = User global search for each tile
+  vw::uint16  seed_option;     /* 0 = User global search for each tile
                                   1 = Narrow search for each tile to low
                                       res disparity seed
                                   2 = Affine Transform and narrow search
